+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+colorscad:
+- Switch -v to enable verbose logging. Decrease verbosity by default, by using OpenSCAD's --quiet parameter.
 
 ## [0.3.1] - 2021-07-03
 ### Fixed